Fixed high memory usage problem
The problem was that graphics objects were being generated each square on each board position, and for each piece on the squares. Adding an additional condition to only generate graphics objects if the squares and pieces are actually part of the board that is being played on (as opposed to boards that the computer is using to think), has solved the problem.